💣Move factory classes in Content namespace to Content.Factories namespace
The ISoundLoader and SoundLoader types were not mean to be public
