+++ Velaptor/Content/Factories/AtlasDataFactory.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.Content
+namespace Velaptor.Content.Factories
 {
     // ReSharper disable RedundantNameQualifier
     using System.Collections.Generic;
