+++ Velaptor/Content/Factories/ITextureFactory.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.Content
+namespace Velaptor.Content.Factories
 {
     // ReSharper disable RedundantNameQualifier
     using Velaptor.Graphics;
@@ -15,13 +15,13 @@ namespace Velaptor.Content
     internal interface ITextureFactory
     {
         /// <summary>
-        /// Creates a <see cref="ITexture"/> object for use for rendering tot he screen.
+        /// Creates an <see cref="ITexture"/> object for use for rendering to the screen.
         /// </summary>
         /// <param name="name">The name of the texture.</param>
         /// <param name="filePath">The file path to the texture.</param>
         /// <param name="imageData">The image data of the texture.</param>
         /// <param name="isPooled">True to have the object pooled.</param>
-        /// <returns>The texture instance.</returns>
+        /// <returns>The <see cref="ITexture"/> instance.</returns>
         ITexture Create(string name, string filePath, ImageData imageData, bool isPooled);
     }
 }
