+++ Velaptor/Content/SoundLoader.cs
@@ -9,6 +9,7 @@ namespace Velaptor.Content
     using System.Collections.Concurrent;
     using System.Diagnostics.CodeAnalysis;
     using System.IO.Abstractions;
+    using Velaptor.Content.Factories;
     using Velaptor.Factories;
 
     // ReSharper restore RedundantNameQualifier
@@ -27,13 +28,11 @@ namespace Velaptor.Content
         /// <summary>
         /// Initializes a new instance of the <see cref="SoundLoader"/> class.
         /// </summary>
-        /// <param name="soundPathResolver">Resolves the path to the sound content.</param>
-        /// <param name="soundFactory">Creates sound instances.</param>
         [ExcludeFromCodeCoverage]
-        public SoundLoader(IPathResolver soundPathResolver, ISoundFactory soundFactory)
+        public SoundLoader()
         {
-            this.soundPathResolver = soundPathResolver;
-            this.soundFactory = soundFactory;
+            this.soundPathResolver = PathResolverFactory.CreateSoundPathResolver();
+            this.soundFactory = IoC.Container.GetInstance<ISoundFactory>();
             this.path = IoC.Container.GetInstance<IPath>();
         }
 
@@ -86,7 +85,7 @@ namespace Velaptor.Content
 
             return this.sounds.GetOrAdd(filePath, (key) =>
             {
-                var sound = this.soundFactory.CreateSound(key);
+                var sound = this.soundFactory.Create(key);
                 sound.IsPooled = true;
 
                 return sound;
